+++ moor/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.2.0
 - Blob data type
+- `insertOrReplace` method for insert statements
 
 ## 1.1.0
 - Transactions